@@ -11,57 +11,58 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.6",
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "~7.3.7",
-    "@angular/common": "~7.2.15",
-    "@angular/compiler": "~7.2.15",
-    "@angular/core": "~7.2.15",
-    "@angular/forms": "~7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.15",
-    "@angular/platform-browser-dynamic": "~7.2.15",
-    "@angular/router": "~7.2.15",
-    "@auth0/angular-jwt": "^2.1.0",
-    "@types/socket.io": "^2.1.2",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.6.9",
-    "express": "^4.17.1",
-    "hammerjs": "^2.0.8",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "~8.2.13",
+    "@angular/compiler": "~8.2.13",
+    "@angular/core": "~8.2.13",
+    "@angular/forms": "~8.2.13",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.13",
+    "@angular/platform-browser-dynamic": "~8.2.13",
+    "@angular/router": "~8.2.13",
+    "@ng-idle/core": "^8.0.0-beta.4",
+    "@ng-idle/keepalive": "^8.0.0-beta.4",
+    "admin-lte": "^3.0.1",
+    "ag-grid-angular": "^22.0.0",
+    "ag-grid-community": "^22.1.1",
+    "angular-jwt": "^0.1.11",
+    "angular2-moment": "^1.9.0",
+    "bootstrap": "^4.4.1",
+    "install": "^0.13.0",
     "jquery": "^3.4.1",
+    "jspdf": "^1.5.3",
     "jwt-decode": "^2.2.0",
-    "moment": "^2.24.0",
-    "ngx-bootstrap": "^4.2.0",
-    "ngx-pagination": "^3.2.1",
-    "ngx-socket-io": "^2.1.1",
-    "ngx-spinner": "^7.1.4",
-    "popper": "^1.0.1",
-    "popper.js": "^1.15.0",
-    "rxjs": "~6.3.3",
-    "socket.io": "^2.2.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "ng-idle": "^1.3.2",
+    "ngx-pagination": "^5.0.0",
+    "ngx-socket-io": "^3.0.1",
+    "ngx-spinner": "^8.1.0",
+    "npm": "^6.13.1",
+    "popper.js": "^1.16.0",
+    "rxjs": "^6.5.3",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@types/googlemaps": "^3.36.3",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "^0.803.20",
+    "@angular/cli": "~8.3.20",
+    "@angular/compiler-cli": "~8.2.13",
+    "@angular/language-service": "~8.2.13",
+    "@types/googlemaps": "^3.38.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^8.10.59",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "tslint": "~5.15.0",
+    "typescript": "^3.4.5"
   }
 }
